requirements.txt
flask
numpy
pandas
python-dotenv
gunicorn
plotly
tweepy
pyrebase
jwt
firebase-admin
textblob
scipy
 matplotlib
 